Move MCP URL generator to transport package
