+++ pkg/client/config.go
@@ -13,8 +13,6 @@ import (
 	"github.com/tailscale/hujson"
 
 	"github.com/stacklok/toolhive/pkg/logger"
-	"github.com/stacklok/toolhive/pkg/transport/ssecommon"
-	"github.com/stacklok/toolhive/pkg/transport/streamable"
 	"github.com/stacklok/toolhive/pkg/transport/types"
 )
 
@@ -458,18 +456,6 @@ func Upsert(cf ConfigFile, name string, url string, transportType string) error 
 		return cf.ConfigUpdater.Upsert(name, MCPServer{Url: url})
 	}
 	return nil
-}
-
-// GenerateMCPServerURL generates the URL for an MCP server
-func GenerateMCPServerURL(transportType string, host string, port int, containerName string) string {
-	// The URL format is: http://host:port/sse#container-name
-	// Both SSE and STDIO transport types use an SSE proxy
-	if transportType == types.TransportTypeSSE.String() || transportType == types.TransportTypeStdio.String() {
-		return fmt.Sprintf("http://%s:%d%s#%s", host, port, ssecommon.HTTPSSEEndpoint, containerName)
-	} else if transportType == types.TransportTypeStreamableHTTP.String() {
-		return fmt.Sprintf("http://%s:%d/%s", host, port, streamable.HTTPStreamableHTTPEndpoint)
-	}
-	return ""
 }
 
 // retrieveConfigFileMetadata retrieves the metadata for client configuration files for a given client type.
